#334: Move & Split AppHelper into helpers folder
