+++ CourageScores/ClientApp/src/helpers/build.test.js
@@ -1,0 +1,7 @@
+// noinspection JSUnresolvedReference
+
+describe('build', () => {
+    describe('getBuildDetail', () => {
+        // tested via App.test.js as it requires access to the DOM
+    });
+});
