+++ CourageScores/ClientApp/src/helpers/errors.js
@@ -24,18 +24,3 @@ export function mapForLogging(error, account) {
         url: window.location.href,
     };
 }
-
-function getBuildDetail(name) {
-    const meta = document.querySelector(`meta[name="build:${name}"]`);
-    return meta
-        ? meta.getAttribute('content')
-        : null;
-}
-
-export function getBuild() {
-    return {
-        branch: getBuildDetail('branch'),
-        version: getBuildDetail('sha'),
-        date: getBuildDetail('date'),
-    };
-}
